Ranking  - Infinity Scroll  - api Update AnimeDetail  - Recommend   - Data Binding   - Navigation   - AnimeRecommends UI Dev  - Series   - Data Binding   - Navigation   - AnimeSeries UI Dev

+++ core/model/src/main/java/com/jparkbro/model/ranking/RankingItem.kt
@@ -14,15 +14,15 @@ data class RankingItem(
     @SerialName("rank")
     val rank: Int,
     @SerialName("change")
-    val change: String? = null,
+    val change: String? = null, // 'N' : "NEW"
     @SerialName("trend")
     val trend: RankingTrend? = null,
     @SerialName("genres")
     val genres: List<String>,
     @SerialName("popularity")
-    val popularity: Long,
+    val popularity: Int,
     @SerialName("trending")
-    val trending: Long? = null,
+    val trending: Int? = null,
 )
 
 @Serializable
